### 2023-10-06 #5818 @AJQuinn
Hi @gkthiruvathukal - I was skimming through the backlog and saw this review waitlisted. I'd be happy to edit if it's still waiting for editorial capacity? I've worked with a few regression packages so it could be a good fit.
