### 2023-10-09 #5818 @AJQuinn
Hi @jeancmaia - I'll be your editor during the review processing. First job is to find some reviewers to take a look over the toolbox and paper.

Could you take a look at the DOIs in the paper - there are few missing/invalid in our automated checking tool. I think two things may be going on. There are a lot more references in your bib than are referenced in the paper, and many of the references don't include an explicit doi within the paper info.

eg within https://github.com/jeancmaia/mcglm/blob/main/paper.bib

```
@article{Jorgensen:2016,
	author={J{\o}rgensen, B. and Kokonendji, C.},
	title={Discrete Dispersion Models and Their Tweedie Asymptotics},
	journal={AStA Advances in Statistical Analysis},
	volume = {100},
	number = {},
	pages={43--78},
	year={2016},
}
```
should include doi line

```
doi = {10.1007/s10182-015-0250-z}
```

you can rerun the reference check to see any changes with the relevant editorialbot line above
