### 2023-10-10 #5818 @AJQuinn
👋 Hello - @eigenfoo, @inakleinbottle, @soodoku and @ejolly would any of you be willing to review this submission for JOSS? The code repository and paper details are further up in this thread.

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html - please check them out for more information and please feel free to ask questions in this thread.

Cheers!
