### 2023-10-10 #5818 @soodoku
hey @AJQuinn, thanks for getting in touch. Unfortunately, I can't take it on. Is there a way to opt out of the JOSS pool of reviewers?
