### 2023-10-11 #5818 @AJQuinn
Hi @soodoku - no problem, thanks for letting me know. I believe you can deselect an ' Available to review' checkbox on your reviewer profile page here: https://reviewers.joss.theoj.org/  Let me know if you run into problems with this. Thanks for your past reviews!
