+++ paper.bib
@@ -1,4 +1,4 @@
-@article{pymc3:2016,
+@article{pymc3:2016,  
 author = {Salvatier, John and Wiecki, Thomas and Fonnesbeck, Christopher},
 year = {2016},
 month = {01},
@@ -12,7 +12,8 @@ author = {Meent, Jan-Willem and Paige, Brooks and Yang, Hongseok and Wood, Frank
 year = {2018},
 month = {09},
 pages = {},
-title = {An Introduction to Probabilistic Programming}
+title = {An Introduction to Probabilistic Programming},
+doi = {10.48550/arXiv.1809.10756}
 }
 
 @article{pyro:2018,
@@ -20,7 +21,8 @@ author = {Bingham, Eli and Chen, Jonathan and Jankowiak, Martin and Obermeyer, F
 year = {2018},
 month = {10},
 pages = {},
-title = {Pyro: Deep Universal Probabilistic Programming}
+title = {Pyro: Deep Universal Probabilistic Programming},
+doi = {10.48550/arXiv.1810.09538}
 }
 
 @article{stan:2017,
@@ -47,33 +49,13 @@ doi = {10.18637/jss.v084.i04}
 }
 
 
-@article{covariance_model:2011,
-author = {Carey VJ, Wang Y},
-year = {2011},
-pages = {3117-3124},
-title = {Working Covariance Model Selection for Generalized Estimating
-Equations.},
-journal = {Statistics in Medicine},
-doi = {110.1002/sim.4300}
-}
-
-@article{godambe:2011,
-author = {Hwang, S and Basawa, I.V},
-year = {2011},
-month = {08},
-pages = {1121-1127},
-title = {Godambe estimating functions and asymptotic optimal inference},
-volume = {81},
-journal = {Statistics and Probablitity Letters},
-doi = {10.1016/j.spl.2011.03.006}
-}
-
 @article{kokonendji:2015,
 author = {Jørgensen, B. and Kokonendji, C. C.},
 year = {2015},
 title = {Discrete dispersion models and their tweedie asymptotics.},
 journal = {AStA Advances in Statistical Analysis, 100(1)},
-pages = {43–78},
+pages = {43-78},
+doi = {10.48550/arXiv.1409.7482}
 }
 
 @article{bent:1987,
@@ -81,6 +63,7 @@ author = {Jørgensen, B.},
 year = {1987},
 title = {Exponential dispersion models.},
 journal = {Journal of the Royal Statistical Society},
+doi = {10.1111/j.2517-6161.1987.tb01685.x}
 }
 
 @article{bent:1997,
@@ -88,78 +71,7 @@ author = {Jørgensen, B.},
 year = {1997},
 title = {The theory of dispersion models.},
 journal = {CRC Press},
-}
-
-@article{covariance_link:1996,
-author = {Chiu TYM, Leonard T, Tsui K },
-year = {1996},
-title = {The Matrix-Logarithmic Covariance Model.},
-journal = {Journal of the American Statistical Association},
-doi = {10.1080/01621459.1996.10476677.}
-}
-
-
-@article{Jennrich:1969,
-    author = {Jennrich, Robert I.},
-    title = "{A Newton-Raphson algorithm for maximum likelihood factor analysis}",
-    journal = {Psychometrika},
-    volume = {34},
-    year = {1969},
-    month = {3},
[…]
