+++ paper.bib
@@ -153,7 +153,7 @@ author = {Galton, F},
 year = {1886},
 pages = {246-263},
 title = {Regression towards mediocrity in hereditary stature.},
-journal = {Journal of the Anthropological Institute of Great Britain and Ireland}
+journal = {Journal of the Anthropological Institute of Great Britain and Ireland},
 doi = {10.2307/2841583}
 }
 
