+++ paper.bib
@@ -154,7 +154,7 @@ year = {1886},
 pages = {246-263},
 title = {Regression towards mediocrity in hereditary stature.},
 journal = {Journal of the Anthropological Institute of Great Britain and Ireland},
-doi = {10.2307/2841583}
+doi = {10.2307/2841583},
 }
 
 
