### 2023-11-08 #6037 @editorialbot
[REVIEW]: Multivariate Covariance Generalized Linear Models in Python: The mcglm library
**Submitting author:** @jeancmaia (<a href="http://orcid.org/0009-0001-3747-0669">Jean Carlos Maia</a>)
**Repository:** https://github.com/jeancmaia/mcglm
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.4
