### 2023-12-01 jeancmaia/mcglm#5 @Spaak
Installation fails via pip (macOS)
Hi, I'm a reviewer for JOSS and am trying to install mcglm via pip (see review thread here https://github.com/openjournals/joss-reviews/issues/6037).

To emulate a typical (using conda), yet still clean (fresh environment), setup, I went:

```
conda create -n reviewmcglm
conda activate reviewmcglm
conda install python=3.9
conda install patsy matplotlib statsmodels scipy numpy pandas seaborn
```

which all went fine. Note that I installed the dependencies here already via conda (the 'default' versions).

Then I went:

```
pip install mcglm
```

which resulted in several errors (see below), which appear to be related to pip trying to compile numpy and failing. This is not something I've encountered before (and have used numpy on many occasions in many different settings), so I suspect it's due to something strange in the dependencies and/or the relation to my machine (an M1 (ARM) mac). I did notice that in mcglm's dependencies you specify exact versions of libraries, rather than accepted ranges. This I suspect is in general undesirable, and installing mcglm should be possible in a typical conda setup like I showed above. A fix here would be appreciated so I can install and continue to review :) (I'm deliberately not installing via git pull since installation is part of the review.)

[Full output of failing pip install](https://github.com/jeancmaia/mcglm/files/13527385/errlog.txt)
