### 2023-12-01 jeancmaia/mcglm#5 @Spaak
I get the same (or similar) errors if I just create a clean conda env, install only python in it (and dependencies, including pip), and then do `pip install mcglm`, so without first installing any further dependencies via conda.
