+++ requirements.txt
@@ -1,7 +1,5 @@
-patsy == 0.5.3
-matplotlib == 3.5.2
-statsmodels == 0.13.2
-scipy == 1.7.1
-numpy == 1.22
-pandas == 1.3.3
-seaborn == 0.11.2
+patsy >= 0.5.2
+matplotlib >= 3.5.2
+statsmodels >= 0.13.2
+pandas >= 1.3.3
+numpy >= 1.25
