### 2023-12-01 jeancmaia/mcglm#5 @jeancmaia
Hey, Spaak. Thank you for reaching out, and sorry for this annoying setback. I have overhauled the library, which looks correct at the moment. Check out the logs:

```
conda create -n reviewmcglm1
source activate reviewmcglm1
pip install mcglm
Collecting mcglm
  Downloading mcglm-0.2.2-py3-none-any.whl (23 kB)
Collecting numpy<2.0,>=1.25
  (18.2 MB)
     |████████████████████████████████| 18.2 MB 1.4 MB/s
Collecting pandas<2.0.0,>=1.3.3
(12.2 MB)
     |████████████████████████████████| 12.2 MB 56 kB/s
Collecting matplotlib<4.0.0,>=3.5.2
 (11.6 MB)
     |████████████████████████████████| 11.6 MB 1.9 MB/s
Collecting statsmodels<0.14.0,>=0.13.2
 (9.9 MB)
Collecting patsy<0.6.0,>=0.5.2
(233 kB)
     |████████████████████████████████| 233 kB 5.0 MB/s
```

```
conda create -n reviewmcglm2
source activate reviewmcglm2
pip install patsy matplotlib statsmodels pandas numpy
pip install mcglm
```

To fix it, I have weeded out unnecessary requirements and polished all library version specifications. Could you try it out again? Thank you so much.
