### 2023-12-07 jeancmaia/mcglm#7 @Spaak
Going through tutorial
(As mentioned before, I'm a reviewer for JOSS, see https://github.com/openjournals/joss-reviews/issues/6037)

I'm going through the tutorial here https://jeancmaia.github.io/posts/tutorial-mcglm/tutorial_mcglm.html and wanted to highlight a few things.

- At the smoking data example, with the simple linear regression, `modelresults.pearson_residuals` is all `NaN`s. At the step before that, you emphasize checking these residuals, so a natural thing would be to also do that here. Likely relevant warning:
```
[/opt/homebrew/Caskroom/miniconda/base/envs/mcglm3/lib/python3.9/site-packages/mcglm/mcglmcattr.py:719](https://file+.vscode-resource.vscode-cdn.net/opt/homebrew/Caskroom/miniconda/base/envs/mcglm3/lib/python3.9/site-packages/mcglm/mcglmcattr.py:719): RuntimeWarning: invalid value encountered in sqrt
  sqrt_mu_power = np.sqrt(mu_power)
[/opt/homebrew/Caskroom/miniconda/base/envs/mcglm3/lib/python3.9/site-packages/mcglm/mcglmcattr.py:724](https://file+.vscode-resource.vscode-cdn.net/opt/homebrew/Caskroom/miniconda/base/envs/mcglm3/lib/python3.9/site-packages/mcglm/mcglmcattr.py:724): RuntimeWarning: invalid value encountered in log
  n=n_len, values=((mu_power * np.log(mu)) [/](https://file+.vscode-resource.vscode-cdn.net/) (2 * sqrt_mu_power))
```
- Is the pAIC metric always the one to look at when comparing different models? (A full treatment of different goodness-of-fit models is beyond the scope, of course, but I'd recommend at least briefly describing and reflecting on pAIC if that's the one you use.) (And what's the difference between AIC (I know that one) and pAIC?)
- With the smoking data example, it would help to explain what is the difference when going to the Tweedie variance model (which you specify in the code); how does that relate to Poisson (which you describe in the text). Also, what is again the difference when going to poisson_tweedie?
- Some more details would be helpful on the construction of the Z matrices in the autoregressive example. What are these, and what is `mc_ma` doing for us exactly? (The documentation of `mc_ma` is not sufficient; also the naming could be a bit more informative.)
- In the last example (soya data), it is not clear how the multivariate model (grain, seeds, and viablePeas as dependent variables) is different from separate univariate models for each of the DVs. I think the primary value in MCGLM models (over existing generalized LM and mixed model approaches) is in properly handling multivariate responses. Therefore, this would be important to highlight also in the tutorial.
