+++ pyproject.toml
@@ -20,6 +20,7 @@ coverage = "^6.4.1"
 jupyter = "^1.0.0"
 pylint = "^2.14.3"
 jupyterlab = "^3.6.1"
+seaborn = "^0.13.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
