### 2023-12-07 9a80716
Changing Default value power to 1
