### 2023-12-07 072516d
New library version
