### 2023-12-08 cc21a96
docstring mc_ma
