+++ tests/test_dependencies.py
@@ -77,39 +77,7 @@ def test_mc_mixed():
             ]
         ),
     )
-    """
-    np.testing.assert_equal(
-        z[1],
-        np.array(
-            [
-                [1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0, 4.0],
-                [2.0, 4.0, 6.0, 8.0, 2.0, 4.0, 6.0, 8.0],
-                [3.0, 6.0, 9.0, 12.0, 3.0, 6.0, 9.0, 12.0],
-                [4.0, 8.0, 12.0, 16.0, 4.0, 8.0, 12.0, 16.0],
-                [1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0, 4.0],
-                [2.0, 4.0, 6.0, 8.0, 2.0, 4.0, 6.0, 8.0],
-                [3.0, 6.0, 9.0, 12.0, 3.0, 6.0, 9.0, 12.0],
-                [4.0, 8.0, 12.0, 16.0, 4.0, 8.0, 12.0, 16.0],
-            ]
-        ),
-    )
 
-    np.testing.assert_equal(
-        z[2],
-        np.array(
-            [
-                [2.0, 3.0, 4.0, 5.0, 2.0, 3.0, 4.0, 5.0],
-                [3.0, 4.0, 5.0, 6.0, 3.0, 4.0, 5.0, 6.0],
-                [4.0, 5.0, 6.0, 7.0, 4.0, 5.0, 6.0, 7.0],
-                [5.0, 6.0, 7.0, 8.0, 5.0, 6.0, 7.0, 8.0],
-                [2.0, 3.0, 4.0, 5.0, 2.0, 3.0, 4.0, 5.0],
-                [3.0, 4.0, 5.0, 6.0, 3.0, 4.0, 5.0, 6.0],
-                [4.0, 5.0, 6.0, 7.0, 4.0, 5.0, 6.0, 7.0],
-                [5.0, 6.0, 7.0, 8.0, 5.0, 6.0, 7.0, 8.0],
-            ]
-        ),
-    )
-    """
 
 
 def test_mc_id():
