### 2023-12-08 jeancmaia/mcglm#7 @jeancmaia
Hey, Spaak. Thanks for the comprehensive feedback. I'll run through each topic according to your sorting list.

1. Nice catch! Owing to a warning suppression,  I couldn't perceive those issues. A smidgen bug wrapped $mu$'s variance calculation when the variance function is constant. The new library version '0.2.3' holds this bugfix.

2. Amazing question! Information Criterion(IC) metrics, such as Akaike and Bayesian, are commonly used for model comparison. They rely on the Likelihood function, which doesn't hold thoroughly for the MCGLM model owing to the second-moment assumptions. Therefore, MCGLM derives goodness-of-fit-like metrics dubbed pseudo-Akaike and Bayesian information, gearing up the Gaussian pseudo log-likelihood function proposed by (Carey and Wang). Picking between pAIC and pBIC is a matter of the problem and its context, as they differ on penalty terms and should be picked accordingly.

3. Tweedie distribution is a fabulous agnostic model that behaves like some exponential family distribution, conditioned by the power parameter. When power is 1, it operates as a Poisson model. https://en.wikipedia.org/wiki/Tweedie_distribution. Tweedie and Poisson Tweedie have different variance functions and behave distinctly; the article to Joss elaborates on it.

4. Z matrices work similarly to the working correlation matrix in Generalized Estimation Equations. They reflect the covariance structure within an outcome variable. The original paper elaborates on how many canonical statistical models are reachable from an adequate specification of Z matrices, such as Moving Average and Mixed models. Auxiliary methods `mc_ma` and `mc_mixed` help one produce those matrices for the statistical analysis. Nevertheless, one can input those matrices as she wishes. I have polished the mc_ma docstring method.

5. The key feature of the multivariate case is the correlation coefficients among outcomes, conveyed in the report's last section: The rho parameters.

We could enhance the Joss article with some of this discussion. Thank you!
