### 2023-12-20 jeancmaia/mcglm#7 @Spaak
Thanks for the clear responses!

Pinging @AJQuinn (JOSS editor) here. I'm not entirely sure how JOSS reviews proceed after something like this. The tutorial is part of the documentation, and I gave some suggestions to improve it. Can I safely move on to the rest of the review checklist, or should I aim to check whether the tutorial and paper have indeed improved after I leave such comments? (I'd prefer moving on and leaving a second/third/etc check to either the editor or a re-review, if it comes to that. Otherwise I'm worried the review process will take ages.) Thanks :)
