### 2023-12-20 jeancmaia/mcglm#10 @Spaak
MCGLM could accept numpy arrays directly (compatibility with statsmodels)
Reviewing for JOSS, https://github.com/openjournals/joss-reviews/issues/6037

You emphasize in the paper/documentation that MCGLM inherits API standards from statsmodels. However, MCGLM requires y/X input to be wrapped as pandas DataFrames, whereas statsmodels accepts both those but also numpy arrays directly. I would recommend accepting numpy arrays to maintain API compatibility with statsmodels. (Note also mc_id etc should then accept arrays, rather than only dataframes.)
