### 2023-12-20 jeancmaia/mcglm#10 @Spaak
I now see that the API docs specify `array_like` as input for endog/exog, so by docs they should accept numpy. However, this results in an error.
