### 2023-12-20 #6037 @Spaak
@jeancmaia As for API docs: some more details could be added. power_fixed/maxiter/tol/tuning/weights are all underspecified. (I can imagine some things they do in the optimization step, but better to make this explicit, e.g. by briefly explaining the type of optimization done here.) Also Ntrial has a capital N (against common style).
