### 2023-12-20 #6037 @Spaak
@jeancmaia I appreciate the extensive tutorial (https://github.com/jeancmaia/mcglm/issues/7) but that appears to be the only web presence besides the PyPI install page (and github). I'd recommend at least making sure there is a hosted version of the API docs available somewhere.

Also there are no community guidelines (see review checklist item) anywhere, I'd recommend adding these to either wherever you end up putting the web API docs or (and) simply in README.md on github.
