### 2023-12-21 df5dcea
Exteding code to manage numpy arrays to
