+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mcglm"
-version = "0.2.3"
+version = "0.2.4"
 description = "Multivariate Covariance Generalized Linear Models"
 authors = ["Jean Carlos Faoot Maia <jeanclmaia@gmail.com>"]
 readme = "README.md"
