### 2023-12-21 jeancmaia/mcglm#10 @jeancmaia
Nice catch! I have tweaked the code, and you have in the version '0.2.4' a library reliable for both pandas and array inputs. Can you check it out again?

Thank you for the great feedback.
