+++ paper.md
@@ -27,109 +27,113 @@ bibliography: paper.bib
 
 # Abstract
 
-Multivariate Covariance Generalized Linear Models (McGLM) are an extension of 
-classic generalized linear models that allow analysis of multivariate and 
-non-independent response variables. Due to its flexibility and explicit 
-specification, McGLM supports many statistical analyses across a wide variety 
-of data and distinct traits. To promote the use of McGLM for statistical analysis, 
-we introduce the `mcglm` library, a new Python library.
+The `mcglm` library, a newly introduced Python tool, facilitates statistical 
+analyses using Multivariate Covariance Generalized Linear Models (McGLM). This 
+contemporary family of models universalizes the traditional Generalized Linear 
+Models(GLM), empowering them to handle multivariate and non-independent response 
+variables. Due to its flexibility and explicit specification, McGLM supports a 
+lot of statistical analyses across different kinds of data and distinct traits. 
+In this article, we promote `mcglm` as a comprehensive framework for statistical 
+analysis in Python. 
 
 The `mcglm` library provides a comprehensive platform for data analysis using 
-the McGLM framework. It inherits the standards of the `statsmodels` library, 
-providing a comprehensive summary report for fitting assessment that includes 
-regression and dispersion coefficients, confidence intervals, hypothesis testing, 
-residual analysis, goodness-of-fit measurements, and correlation coefficients 
-between outcomes. In addition, the library provides a rich set of link and 
-variance functions and tools to define inner-response dependency matrices 
-through the matrix linear predictor. The base code is both extendable and 
-reliable thanks to good object-oriented programming practices and unit testing.
+the McGLM framework. Built upon the established standards of the `statsmodels` 
+library, it provides a comprehensive summary report for fitting assessment, 
+including crucial elements such as regression and dispersion coefficients, 
+confidence intervals, hypothesis testing results, residual analysis, 
+goodness-of-fit measurements, and correlation coefficients between outcomes. 
+In addition, the library provides a rich set of link and variance functions 
+and tools to define inner-response dependency matrices through the matrix 
+linear predictor. The base code is extendable and reliable, reflecting sound 
+object-oriented programming practices and thorough unit testing. 
 
-The library is hosted on PyPI and can be installed with the aid of some Python 
-library manager, such as `pip`.
+The library is hosted on PyPI and can be installed with some Python library 
+manager, such as `pip`.
 
 # Introduction
 
-Dated at the beginning of the 19th century and controversial about actual 
+Dated at the beginning of the 19th century and controversial about the actual 
 authorship, the least squares method established an optimization algorithm 
-[@10.1214/aos/1176345451]. According to the Gauss-Markov theorem 
-[@Gauss-Marc], the resulting estimates are optimal and unbiased under 
-linear conditions. This optimization method forms the basis of linear 
-regression, one of the earliest statistical models [@galton:1886; @linearregression:1982]. 
-Linear regression associates a response variable to a group of covariates 
-by employing a linear operation on regression coefficients 
-[@10.2307/2333849]. Three main assumptions for linear regression are 
-linearity, independent realizations of the response variable, and a Gaussian 
-homoscedastic error with a zero mean. In the subsequent years, many other 
-proposals have aimed to generalize these assumptions; however, the least 
-squares method and its application in linear regression still significantly 
-impact statistical modeling and data analysis. 
[…]
