### 2023-12-26 d466356
Refactoring MCGLM documentation
