### 2023-12-26 b314948
Weeding out doc files
