+++ docs/source/index.rst
@@ -1,20 +1,0 @@
-.. mcglm documentation master file, created by
-   sphinx-quickstart on Wed Aug  3 08:36:42 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to mcglm's documentation!
-=================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
