+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. mcglm documentation master file, created by
+   sphinx-quickstart on Tue Dec 26 17:43:55 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to mcglm's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
