+++ docs/requirements.txt
@@ -1,0 +1,1 @@
+sphinx-rtd-theme >= 2.0
