+++ docs/source/conf.py
@@ -6,6 +6,10 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../mcglm/'))
+
 project = 'mcglm'
 copyright = '2023, Jean Carlos Faoot Maia'
 author = 'Jean Carlos Faoot Maia'
@@ -14,7 +18,9 @@ release = '0.2.4'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+extensions = ['sphinx.ext.autodoc',
+              'sphinx.ext.viewcode',
+              'sphinx.ext.napoleon']
 
 templates_path = ['_templates']
 exclude_patterns = []
@@ -26,8 +32,3 @@ exclude_patterns = []
 
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
-
-
-import os
-import sys
-sys.path.insert(0, os.path.abspath('../../mcglm/'))
