+++ docs/requirements.txt
@@ -1,6 +1,2 @@
 sphinx-rtd-theme >= 2.0
-patsy >= 0.5.2
-matplotlib >= 3.5.2
-statsmodels >= 0.13.2
-pandas >= 1.3.3
-numpy >= 1.25
+mcglm >= 0.2.3
