+++ docs/source/conf.py
@@ -8,7 +8,7 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../mcglm'))
+sys.path.insert(0, os.path.abspath('../..'))
 
 project = 'mcglm'
 copyright = '2023, Jean Carlos Faoot Maia'
