+++ docs/html/_sources/mcglm.rst.txt
@@ -1,0 +1,63 @@
+mcglm library
+=============
+
+mcglm.mcglm module
+------------------
+
+.. automodule:: mcglm.mcglm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Submodules
+----------
+
+mcglm.dependencies module
+-------------------------
+
+.. automodule:: mcglm.dependencies
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+mcglm.mcglmcattr module
+-----------------------
+
+.. automodule:: mcglm.mcglmcattr
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+mcglm.mcglmmean module
+----------------------
+
+.. automodule:: mcglm.mcglmmean
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+mcglm.mcglmvariance module
+--------------------------
+
+.. automodule:: mcglm.mcglmvariance
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+mcglm.utils module
+------------------
+
+.. automodule:: mcglm.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: mcglm
+   :members:
+   :undoc-members:
+   :show-inheritance:
