+++ docs/html/_sources/modules.rst.txt
@@ -1,7 +1,0 @@
-mcglm
-=====
-
-.. toctree::
-   :maxdepth: 4
-
-   mcglm
