+++ mcglm/mcglm.py
@@ -21,32 +21,58 @@ class MCGLM(MCGLMMean, MCGLMVariance):
     __doc__ = """
     MCGLM class that implements MCGLM stastical models. (Bonat, Jørgensen 2015)
         
-    It extends GLM for multi-responses and dependent components by fitting second-moment assumptions.
+    It extends GLM for multi-responses and dependent components by fitting 
+    second-moment assumptions.
 
     Parameters
     ----------
     endog : array_like
-        1d array of endogenous response variable. In case of multiple responses, the user must pass the responses on a list.  
+        1d array of endogenous response variable. In case of multiple 
+        responses, the user must pass the responses on a list.  
     exog : array_like
-        A dataset with the endogenous matrix in a Numpy fashion. Since the library doesn't set an intercept by default, the user must add it. In the case of multiple responses, the user must pass the design matrices as a python list. 
+        A dataset with the endogenous matrix in a Numpy fashion. Since the 
+        library doesn't set an intercept by default, the user must add it. In 
+        the case of multiple responses, the user must pass the design matrices 
+        as a python list. 
     z : array_like
         List with matrices components of the linear covariance matrix.
     link : array_like, string or None
-        Specification for the link function. The MCGLM library implements the following options: identity, logit, power, log, probit, cauchy, cloglog, loglog, negativebinomial. In the case of None, the library chooses the identity link. In multiple responses, user must pass values as list.  
+        Specification for the link function. The MCGLM library implements the 
+        following options: identity, logit, power, log, probit, cauchy, 
+        cloglog, loglog, negativebinomial. In the case of None, the library 
+        chooses the identity link. In multiple responses, user must pass 
+        values as list.  
     variance : array_like, string or None
-        Specification for the variance function. The MCGLM library implements the following options: constant, tweedie, binomialP, binomialPQ, geom_tweedie, poisson_tweedie. In the case of None, the library chooses the constant link. In multiple responses, user must pass values as list.   
+        Specification for the variance function. The MCGLM library implements 
+        the following options: constant, tweedie, binomialP, binomialPQ, 
+        geom_tweedie, poisson_tweedie. In the case of None, the library 
+        chooses the constant link. In multiple responses, user must pass 
+        values as list.   
     offset : array_like or None
-        Offset for continuous or count. In multiple responses, user must pass values as list.   
+        Offset for continuous or count. In multiple responses, user must pass 
+        values as list.   
     ntrial : array_like or None
-        The parameter ntrial is key for binomial responses. In multiple responses, the user must pass values as a list.
+        The parameter ntrial is key for binomial responses. In multiple 
+        responses, the user must pass values as a list.
     power_fixed : array_like or None
-        The parameter power is key for Tweedie-like distributions, as it defines the overall behavior of the model. The library mcglm can also estimate the power parameter if power_fixed equals True. Therefore, in the case of variance functions, either tweedie, geom_tweedie, or poisson_tweedie. In multiple responses, the user must pass values as a list.
+        The parameter power is key for Tweedie-like distributions, as it 
+        defines the overall behavior of the model. The library mcglm can also 
+        estimate the power parameter if power_fixed equals True. Therefore, in 
+        the case of variance functions, either tweedie, geom_tweedie, or 
+        poisson_tweedie. In multiple responses, the user must pass values as a 
+        list.
     maxiter : float or None
-        The parameter maxiter defines the total maximum possible cycles of iterations for the optimization process. Defaults to 200.
[…]
