+++ mcglm/mcglmcattr.py
@@ -18,14 +18,23 @@ from .utils import (
 
 class MCGLMCAttributes:
     """
-    The class "MCGLMCAttributes" has the sake of calculating every C operations, used on throughout adjustments of mean and variance. This class has two interfaces, "c_inverse" and "c_complete"; one for each of two adjustment steps of MCGLM.
+    The class "MCGLMCAttributes" has the sake of calculating every C 
+    operations, used on throughout adjustments of mean and variance. This 
+    class has two interfaces, "c_inverse" and "c_complete"; one for each of 
+    two adjustment steps of MCGLM.
 
-    The interface "c_inverse" crafts only inverse C, and the "c_complete" adds its derivatives and other features onto response. A Quasi-likelihood estimation needs only the inverse of "C" matrix. Therefore c_inverse saves computational resources by avoiding unnecessary operations on mean step adjustment.
+    The interface "c_inverse" crafts only inverse C, and the "c_complete" adds 
+    its derivatives and other features onto response. A Quasi-likelihood 
+    estimation needs only the inverse of "C" matrix. Therefore c_inverse saves 
+    computational resources by avoiding unnecessary operations on mean step 
+    adjustment.
     """
 
     def c_inverse(self, mu, power, rho, tau, full_response=False):
         """
-        A method to generate only the inverse of the C matrix, explicitly made for the mean treatment step. This method interacts with sigma and omega amenities by list of each parameter.
+        A method to generate only the inverse of the C matrix, explicitly made 
+        for the mean treatment step. This method interacts with sigma and 
+        omega amenities by list of each parameter.
 
         Parameters
         ----------
@@ -46,7 +55,9 @@ class MCGLMCAttributes:
 
     def c_complete(self, mu, power, rho, tau):
         """
-        A method to generate the whole list of C components, explicitly made for the variance treatment step. This method interacts with sigma and omega crafting practices, passing the list of each parameter.
+        A method to generate the whole list of C components, explicitly made 
+        for the variance treatment step. This method interacts with sigma and 
+        omega crafting practices, passing the list of each parameter.
 
         Parameters
         ----------
@@ -100,8 +111,16 @@ class MCGLMCAttributes:
         return c_inverse, c_derivatives, c_values
 
     def __generate_c_inverse(self, mu, power, rho, tau, full_response=False):
-        """This method generates and retrieves the inverse of the C matrix, which is a pivotal component of the quasi-score calculation. Notwithstanding the mentioned, this method crafts many important artifacts throughout the time as omega, sigma, sigma Cholesky, the inverse of Cholesky sigma, sigma among responses, sigma block diagonal matrix alongside its inverse.
-        Owing to the previously mentioned worthy artifacts crafted by the method, it can retrieve either only the c inverse matrix or all artifacts. The former is helpful for the method _c_inverse, whereas the former is for _c_complete.
+        """This method generates and retrieves the inverse of the C matrix, 
+        which is a pivotal component of the quasi-score calculation. 
+        Notwithstanding the mentioned, this method crafts many important 
+        artifacts throughout the time as omega, sigma, sigma Cholesky, the 
+        inverse of Cholesky sigma, sigma among responses, sigma block diagonal 
+        matrix alongside its inverse.
+        Owing to the previously mentioned worthy artifacts crafted by the 
+        method, it can retrieve either only the c inverse matrix or all 
+        artifacts. The former is helpful for the method _c_inverse, whereas 
+        the former is for _c_complete.
 
         Parameters
         ----------
@@ -215,7 +234,8 @@ class MCGLMCAttributes:
             sigma_chol_block_matrix : array_like
                 A block diagonal matrix with all Sigmas.
[…]
