+++ mcglm/dependencies.py
@@ -1,5 +1,6 @@
 """
-An extension of MCGLM library to provide three options for matrix linear predictor: mc_id, mc_ma, and mc_mixed.
+An extension of MCGLM library to provide three options for matrix linear 
+predictor: mc_id, mc_ma, and mc_mixed.
 """
 import numpy as np
 import itertools
@@ -11,7 +12,8 @@ from itertools import combinations
 
 def mc_id(data=None):
     """
-    mc_id method retrieves a numpy diagonal matrix with data length of the original matrix
+    mc_id method retrieves a numpy diagonal matrix with data length of the 
+    original matrix
     """
     size = data.shape[0]
     return diagonal(size, np.ones(size))
@@ -19,10 +21,13 @@ def mc_id(data=None):
 
 def mc_ma(id=None, time=None, data=None, order=1):
     """
-    mc_ma method retrieves the Z components for matrix linear predictor associated with Autoregressive models(Feller, W. (1957). An introduction to probability theory and its applications / William Feller.
+    mc_ma method retrieves the Z components for matrix linear predictor 
+    associated with Autoregressive models(Feller, W. (1957). An introduction 
+    to probability theory and its applications / William Feller.
     Wiley New York, 2nd ed. edition.).
 
-    To ilustrate, in a three-row example, a MA(1) produce the following dependence matrix:
+    To ilustrate, in a three-row example, a MA(1) produce the following 
+    dependence matrix:
 
     [[0, 1, 0],
     [1, 0, 1],
@@ -70,7 +75,9 @@ def mc_ma(id=None, time=None, data=None, order=1):
 
 def mc_mixed(data=None, formula=None):
     """
-    mc_mixed retrieves the components for matrix linear predictor associated with mixed models(Demidenko E (2013). Mixed Models: Theory and Applications with R. John Wiley & Sons. doi:10.1002/0471728438.).
+    mc_mixed retrieves the components for matrix linear predictor associated 
+    with mixed models(Demidenko E (2013). Mixed Models: Theory and 
+    Applications with R. John Wiley & Sons. doi:10.1002/0471728438.).
     """
     design_matrix = dmatrix(formula, data=data, return_type="dataframe")
     val_columns = design_matrix.columns.tolist()
