+++ mcglm/mcglmmean.py
@@ -34,27 +34,27 @@ AVAILABLE_LINK_FUNCTIONS = {
 
 class MCGLMMean(MCGLMCAttributes):
     """
-    MCGLMMean is the class for the first moment adjustment within MCGLM 
-    inference. It handles lifecycle completely, ranging from mu and 
+    MCGLMMean is the class for the first moment adjustment within MCGLM
+    inference. It handles lifecycle completely, ranging from mu and
     derivatives to quasi-likelihood calculations.
-    This class has two interfaces: 'calculate_mean_features' which calculates 
-    mu attributes, and 'update_beta' that applies quasi-likelihood estimation 
+    This class has two interfaces: 'calculate_mean_features' which calculates
+    mu attributes, and 'update_beta' that applies quasi-likelihood estimation
     and retrieves a new beta.
-    This class implements the Estimating Equation Quasi-score (Wedderburn, 
+    This class implements the Estimating Equation Quasi-score (Wedderburn,
     1974) and the second-order optimization algorithm (Jennrich, 1969) and
     (Widyaningsih et al., 2017).
 
     References
     ----------
-    Wedderburn, R. W. M. (1974). Quasi-likelihood functions, generalized 
+    Wedderburn, R. W. M. (1974). Quasi-likelihood functions, generalized
     linear models, and the Gauss—Newton method. Biometrika, 61(3):439–447.
 
-    Jennrich, R. I. (1969). A Newton-Raphson algorithm for maximum likelihood 
+    Jennrich, R. I. (1969). A Newton-Raphson algorithm for maximum likelihood
     factor analysis. Psychometrika, 34.
 
-    Widyaningsih, P., Saputro, D. e Putri, A. (2017). Fisher scoring method 
-    for parameter estimation of geographically weighted ordinal logistic 
-    regression (gwolr) model. Journal of Physics: Conference Series, 
+    Widyaningsih, P., Saputro, D. e Putri, A. (2017). Fisher scoring method
+    for parameter estimation of geographically weighted ordinal logistic
+    regression (gwolr) model. Journal of Physics: Conference Series,
     855:012060.
     """
 
@@ -70,7 +70,7 @@ class MCGLMMean(MCGLMCAttributes):
                 A link function
         Returns
         -------
-            statsmodels.genmod.families.links : a corresponding object for the 
+            statsmodels.genmod.families.links : a corresponding object for the
             link function.
         """
         assert link in AVAILABLE_LINK_FUNCTIONS, (
@@ -79,7 +79,7 @@ class MCGLMMean(MCGLMCAttributes):
         return AVAILABLE_LINK_FUNCTIONS.get(link.lower())
 
     def __linear_predictor(self, X, beta):
-        """Method linear predictor applies linear operation between covariates 
+        """Method linear predictor applies linear operation between covariates
         and the regression parameters.
 
         Parameters
@@ -118,7 +118,7 @@ class MCGLMMean(MCGLMCAttributes):
     def __link_function_attributes(
         self, link: str, beta: np.array, X: np.array, offset: int = 0
     ):
-        """The method __link_function_attributes calculates the vector of 
+        """The method __link_function_attributes calculates the vector of
         expected values and its derivatives. It returns data as dictionary
 
         Parameters
@@ -178,7 +178,7 @@ class MCGLMMean(MCGLMCAttributes):
         return mu_attributes_per_response, mu, d
 
     def update_beta(self, beta, W, power, rho, tau):
-        """The method update_beta takes the current beta, leverages the 
+        """The method update_beta takes the current beta, leverages the
         quasi-likelihood estimator to calculate the next regression parameters.
 
         Parameters
@@ -195,7 +195,7 @@ class MCGLMMean(MCGLMCAttributes):
                 Dispersion parameters.
         Returns
[…]

[…]
