### 2023-12-29 #6037 @jeancmaia
@Spaak Thanks for the comprehensive feedback. I have revamped the project by incorporating all the aforementioned items.

The new API Docs is hosted on: https://mcglm.readthedocs.io/en/latest/
The new paper is available on: https://github.com/jeancmaia/mcglm/blob/main/paper.md

It would be great if you could check the project out again. Thanks :)
