### 2024-05-07 #6037 @jeancmaia
Hey @crvernon. I have addressed all the great feedback provided by @Spaak and refactored both the library and the paper.

Thanks!
