### 2024-05-14 #6037 @Spaak
@crvernon The only remaining feedback I have is that the tutorial and paper could contain a little bit more motivating background on why an inference with an MCGLM model is more powerful (in some sense) than something univariate/classical. But I think that is also a matter of preference (related to the journal and its scope); the software itself is adequately documented. All other points have been taken care of!
