### 2024-06-24 #6037 @bkrayfield
@crvernon After reviewing the issues raised by @Spaak and the repo, I was able to check every item off the checklist. I did want to refer back to the community guidelines point of @Spaak. I was able to find them in the docs, but I would recommend taking them out as a separate point and adding a little more detail (more on the type of pull requests the maintainer is looking for, etc...).
