### 2024-06-26 #6037 @crvernon
Thanks @bkrayfield  and  @Spaak!  It looks like your checklists are completed now.

:wave: @jeancmaia I'll take a look now and see if we can get this moving towards publication.

General:
- [ ] To address the need for visibility of the community guidelines, please add in the link to your new docs (https://mcglm.readthedocs.io/) as your docs website in the "About" panel in your GitHub repository.  Currently, users have to way to getting to this easily.

- [ ] LINE 56: "The foremost library statistical analysis..." should be "The foremost library for statistical analysis..."
- [ ] LINES 94-96: please fix the hyperlink referencing here.
- [ ] LIENS 97-98:  Restate "...programming:  The NumPy (Harris, 2020), scipy (Virtanen, 2020), and scipy.sparse." to be "...programming:  NumPy (Harris, 2020) and scipy (Virtanen, 2020)."
- [ ] LINE 105:  Change "For in-depth details, access the code-base https://github.com/jeancmaia/mcglm." to "For in-depth details, access our documentation at https://mcglm.readthedocs.io."
- [ ] Remove LINES 134-135, these are stated throughout.
- [ ] Remove the Discussion section as it has been established in the statement of need and throughout the paper.  We try to keep the paper part of JOSS submissions minimal.
- [ ] Please ensure that all proper nouns have the correct capitalization in the bibliography.  You ensure this in the bib file by wrapping a correct capitalization with curly brackets to maintain their intention once compiled.  An example would be LINE 206 in which "gauss" should be "Gauss"
