### 2024-06-26 b12a999
incorporating crvernon suggestions
