+++ paper.bib
@@ -3,7 +3,7 @@ author = {Salvatier, John and Wiecki, Thomas and Fonnesbeck, Christopher},
 year = {2016},
 month = {01},
 pages = {},
-title = {Probabilistic programming in Python using PyMC3},
+title = {Probabilistic programming in {P}ython using PyMC3},
 doi = {10.7287/PEERJ.PREPRINTS.1686V1}
 }
 
@@ -52,7 +52,7 @@ doi = {10.18637/jss.v084.i04}
 @article{kokonendji:2015,
 author = {Jørgensen, B. and Kokonendji, C. C.},
 year = {2015},
-title = {Discrete dispersion models and their tweedie asymptotics.},
+title = {Discrete dispersion models and their {T}weedie asymptotics.},
 journal = {AStA Advances in Statistical Analysis, 100(1)},
 pages = {43-78},
 doi = {10.48550/arXiv.1409.7482}
@@ -86,7 +86,7 @@ doi = {10.1002/1097-0258(20000730)19:14<1952::AID-SIM474>3.0.CO;2-K}
 }
 
 @article{Krupskii:2013,
-    title = {Factor Copula Models for Multivariate Data},
+    title = {Factor {C}opula {M}odels for Multivariate Data},
     author = {Krupskii, P. and Joe, H.},
     journal = {Journal of Multivariate Analysis},
     volume = {120},
@@ -100,7 +100,7 @@ doi = {10.1002/1097-0258(20000730)19:14<1952::AID-SIM474>3.0.CO;2-K}
 author = {Nelder, J.A. and Wedderburn, R.W.M.},
 year = {1972},
 pages = {370-384},
-title = {Generalized Linear Models.},
+title = {Generalized {L}inear {M}odels.},
 isbn = {978-0-8176-4810-7},
 doi = {10.2307/2344614} 
 }
@@ -110,13 +110,13 @@ author = {Müller, Marlene},
 year = {2004},
 month = {02},
 pages = {},
-title = {Generalized Linear Models},
+title = {Generalized {L}inear {M}odels},
 isbn = {978-3-540-66207-5},
 doi = {10.1007/978-3-642-21551-3_24}
 }
 
 @article{GAM:1986,
-title = {Generalized additive models (with discussion)},
+title = {Generalized {A}dditive {M}odels (with discussion)},
 author = {Hastie, T. and R. Tibshirani },
 journal = {Statistical Science 1},
 pages = {297-318},
@@ -143,7 +143,7 @@ author = {Hallin, Marc},
 year = {2014},
 month = {09},
 pages = {},
-title = {Gauss-Markov Theorem in Statistics},
+title = {Gauss-{M}arkov Theorem in {S}tatistics},
 isbn = {9781118445112},
 doi = {10.1002/9781118445112.stat07536}
 }
@@ -167,7 +167,7 @@ doi = {10.2307/2841583},
  number = {1/2},
  pages = {1--24},
  publisher = {[Oxford University Press, Biometrika Trust]},
- title = {Studies in the History of Probability and Statistics. XV: The Historical Development of the Gauss Linear Model},
+ title = {Studies in the History of Probability and Statistics. XV: The Historical Development of the {G}auss Linear Model},
  urldate = {2022-12-22},
  volume = {54},
  year = {1967},
@@ -220,7 +220,7 @@ author = {Bonat, Wagner},
 year = {2018},
 month = {04},
 pages = {},
-title = {Multiple Response Variables Regression Models in R : The mcglm Package},
+title = {Multiple Response Variables Regression Models in {R} : The mcglm Package},
 volume = {84},
 journal = {Journal of Statistical Software},
 doi = {10.18637/jss.v084.i04}
[…]
