+++ paper.bib
@@ -126,7 +126,7 @@ doi = {10.1214/ss/1177013604}
 
 @article{10.1214/aos/1176345451,
 author = {Stephen M. Stigler},
-title = {{Gauss and the Invention of Least Squares}},
+title = {{G}auss and the invention of {L}east {S}quares},
 volume = {9},
 journal = {The Annals of Statistics},
 number = {3},
@@ -238,15 +238,20 @@ doi = {10.25080/Majora-92bf1922-011}
 }
 
 
-@article{Madasu:2015,
-    author = {Madasu, Vamsi and Venna, Trinadh and Eltaeib, Tarik},
-    year = {2015},
-    month = {03},
-    pages = {3159-40},
-    title = {SOLID Principles in Software Architecture and Introduction to RESM Concept in OOP},
-    volume = {2},
-    journal = {Journal of Engineering Science and Technology},
-    doi = {}
+@article {10.3844/jcssp.2021.624.638,
+article_type = {journal},
+title = {Impact of Design Principles and Patterns on Software Flexibility: An Experimental Evaluation Using Flexible Point (FXP)},
+author = {Rana, Muhammad Ehsan and Khonica, Eddy},
+volume = {17},
+number = {7},
+year = {2021},
+month = {Jul},
+pages = {624-638},
+doi = {10.3844/jcssp.2021.624.638},
+url = {https://thescipub.com/abstract/jcssp.2021.624.638},
+abstract = {Software flexibility is a crucial factor in designing and developing software as it reflects its capability to adapt to changes. It is a topic that has been discussed for a very long time which indicates its importance in software development. However, it is not easy to produce a flexible software design. Design principles provide fundamental concepts in designing good software. On the other hand, design patterns are proven solution to recurring problem. When used correctly, design principles and patterns can be used to improve software flexibility. However, it is necessary to evaluate its impact on software flexibility. For this purpose, this research will conduct an experiment by developing a simple application using Object-Oriented Programming (OOP) based on a case study. Based on the requirements of the case study, two SOLID design principles are chosen such as Single Responsibility Principle (SRP) and Open Closed Principle (OCP) while Strategy and Decorator for the patterns from the GoF. Then, Flexible Points (FXP) is used to measure the differences in software flexibility by comparing the solutions developed before and after applying design principles and patterns. This is aimed to prove that the chosen design principles and patterns have positive impact on software flexibility. Lastly, the result analysis shows that the use of the chosen design principles and patterns indeed improve the flexibility of the software. Therefore, the authors highly recommend adopting design principles and patterns in software development.},
+journal = {Journal of Computer Science},
+publisher = {Science Publications}
 }
 
 
@@ -272,7 +277,7 @@ doi = {10.25080/Majora-92bf1922-011}
 }
 
 @article{Krupskii:2013,
-    title = {Factor Copula Models for Multivariate Data},
+    title = {Factor {C}opula {M}odels for multivariate data},
     author = {Krupskii, P. and Joe, H.},
     journal = {Journal of Multivariate Analysis},
     volume = {120},
