### 2024-06-26 #6037 @jeancmaia
Hi, @crvernon. Thanks for your feedback. I have addressed your list thoroughly and also solved a missing DOI.
