### 2024-06-27 #6037 @crvernon
Thanks @jeancmaia, it looks great.

👋  NAME - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Please respond with the DOI of the archived version and the version number of the release here
