### 2024-06-27 #6037 @jeancmaia
Hi @crvernon. Thanks for helping me throughout this process.

I released a software version named "v0.2.4" and published it at Zenodo. All metadata match appropriately.

Archived version's DOI is: [10.5281/zenodo.12569049](https://zenodo.org/doi/10.5281/zenodo.12569049)

Thank you very much!
