### 2024-06-27 #6037 @crvernon
Thanks @jeancmaia!  A couple of simple things that need to be adjusted:

- [ ] The version number in your software (0.2.1) is different than the version number in your Zenodo archive (v1). Please edit the metadata in the Zenodo archive to represent the actual version of the software.
- [ ] The author order in your Zenodo archive is different than what is in your paper.  I know that these are equal contribution but the order of listing should still be the same.  You can edit this in the metadata as well.

Thanks!
