### 2024-06-27 #6037 @jeancmaia
Thanks @crvernon. Sorry for those minor bugs. Both versions of Zenodo and the library match now; check out those urls: https://zenodo.org/records/12569050, https://github.com/jeancmaia/mcglm/tree/v0.2.4, https://pypi.org/project/mcglm/. Also, authors were rearranged correctly. Can you check it out again? Thanks!
