+++ paper.bib
@@ -37,17 +37,6 @@ doi = {10.18637/jss.v076.i01}
 }
 
 
-@article{Bonat:2018,
-author = {Bonat, Wagner},
-year = {2018},
-month = {04},
-pages = {},
-title = {Multiple Response Variables Regression Models in R : The mcglm Package},
-volume = {84},
-journal = {Journal of Statistical Software},
-doi = {10.18637/jss.v084.i04}
-}
-
 
 @article{kokonendji:2015,
 author = {Jørgensen, B. and Kokonendji, C. C.},
@@ -74,27 +63,6 @@ journal = {CRC Press},
 doi = {10.1002/1097-0258(20000730)19:14<1952::AID-SIM474>3.0.CO;2-K}
 }
 
-
-@article{Masarotto:2012,
-   title = {Gaussian Copula Marginal Regression},
-   author = {Masarotto, G. and Varin, C.},
-   journal = {Electronic Journal of Statistics},
-   volume = {6},
-   pages = {1517--1549},
-   year = {2012},
-   doi = {10.1214/12-EJS721}
-}
-
-@article{Krupskii:2013,
-    title = {Factor {C}opula {M}odels for Multivariate Data},
-    author = {Krupskii, P. and Joe, H.},
-    journal = {Journal of Multivariate Analysis},
-    volume = {120},
-    number = {1},
-    pages = {85--101},
-    year = {2013},
-    doi = {10.1016/j.jmva.2013.05.001}
-}
 
 @article{glm:1972,
 author = {Nelder, J.A. and Wedderburn, R.W.M.},
