### 2024-06-27 #6037 @jeancmaia
I have found duplicated refs on the paper.bib. Solved.
