+++ paper.bib
@@ -143,7 +143,7 @@ doi = {10.2307/2841583},
 }
 
 
-@Article{ linearregression:1982,
+@Article{linearregression:1982,
  title         = {The Minimum Sum of Absolute Errors Regression: A State of the Art Survey.},
  author        = {Narula, Subhash C., and John F. Wellington.},
  year          = {1982},
