### 2024-06-27 aaa15fb
Updating tab:methods
