+++ paper.bib
@@ -92,7 +92,7 @@ year = {1986},
 doi = {10.1214/ss/1177013604}
 }
 
-@article{10.1214/aos/1176345451,
+@article{10.1214/1176345451,
 author = {Stephen M. Stigler},
 title = {{G}auss and the invention of {L}east {S}quares},
 volume = {9},
@@ -126,7 +126,7 @@ doi = {10.2307/2841583},
 }
 
 
-@article{10.2307/2333849,
+@article{gauss:2022,
  ISSN = {00063444},
  URL = {http://www.jstor.org/stable/2333849},
  abstract = {The linear regression model owes so much to Gauss that we believe it should bear his name. Other authors who made substantial contributions are: Cauchy who introduced the idea of orthogonality; Chebyshev who applied it to polynomial models; Pizzetti who found the distribution of the sum of squares of the residuals on the Normal assumption; Karl Pearson who linked the model with the multivariate Normal thereby broadening the field of applications; and R. A. Fisher whose extension of orthogonality to qualitative comparisons laid the foundations of the modern theory of experimental design.},
@@ -206,7 +206,7 @@ doi = {10.25080/Majora-92bf1922-011}
 }
 
 
-@article {10.3844/jcssp.2021.624.638,
+@article {solid:2021,
 article_type = {journal},
 title = {Impact of Design Principles and Patterns on Software Flexibility: An Experimental Evaluation Using Flexible Point (FXP)},
 author = {Rana, Muhammad Ehsan and Khonica, Eddy},
@@ -224,7 +224,7 @@ publisher = {Science Publications}
 
 
 @article{Verbeke:2014,
-  title = {The Analysis of Multivariate Longitudinal Data: A Review},
+  title = {The Analysis of {M}ultivariate {L}ongitudinal Data: A Review},
   author = {Verbeke, G. and Fieuws, S. and Molenberghs, G. and Davidian, M.},
   journal = {Statistical Methods in Medical Research},
   volume = {23},
