### 2024-06-27 #6037 @jeancmaia
That's fabulous. Thanks @crvernon.

For some reason, the paper is not rendering on the JOSS page. https://joss.theoj.org/papers/10.21105/joss.06037. Also, when I attempted to download it, I received a 404. Are you facing this issue too?
